# Add filtering sections by name