--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "ion"
-version = "0.8.0"
+version = "0.8.1"
 description = "parse and process ion files"
 license = "MIT"
 homepage = "http://github.com/pzol/ion_rs"