Existing API is not affected. New API gives performance improvement.

### Benches
Use case: loading ION with 3000 lines, only 2 small sections are actually needed (`CONTRACT` and `DEF.HOTEL`).

```
Without filtering (old API), needed sections on start of ION: 4,187,092 ns/iter (+/- 125,166)
Without filtering (old API), needed sections on end   of ION: 4,223,583 ns/iter (+/- 155,651)
With    filtering (new API), needed sections on start of ION:    15,027 ns/iter (+/- 1,612)
With    filtering (new API), needed sections on end   of ION:   962,318 ns/iter (+/- 31,853)
```